refactor: move copyToClipboard function to sharedUtils and update imports
