+++ packages/vscode-extension/src/network/utils/sharedUtils.ts
@@ -1,0 +1,8 @@
+export async function copyToClipboard(text: string): Promise<void> {
+  try {
+    await navigator.clipboard.writeText(text);
+  } catch (error) {
+    console.error("Failed to copy to clipboard:", error);
+    throw error;
+  }
+}
